Moved raffle form submit to api route to prepare for submit via AJAX
