+++ app/routes/api.py
@@ -1,7 +1,17 @@
-from flask import Blueprint, session, jsonify, abort, request, url_for
+from flask import (
+    abort,
+    Blueprint,
+    current_app,
+    jsonify,
+    redirect,
+    request,
+    session,
+    url_for
+)
 from app.util import reddit
 from app.extensions import rq
 from app.db.models import Raffle
+from app.jobs.raffle_job import raffle
 
 
 api = Blueprint('api', __name__)
@@ -52,8 +62,82 @@ def status():
     return jsonify({'status': status, 'error': job.meta.get('error')})
 
 
+@api.route('/raffles/new', methods=['POST'])
+def new_raffle():
+    form = request.form.copy()
+    if 'submissionUrl' in form:
+        form['submissionUrl'] = _ensure_protocol(form['submissionUrl'])
+    if not _validate_raffle_form(form):
+        current_app.logger.error('Form validation failed {}'.format(
+            [(key, value) for key, value in request.form.items()]))
+        abort(422)
+
+    sub_id = reddit.submission_id_from_url(form.get('submissionUrl'))
+    if _raffle_exists(sub_id):
+        return redirect(url_for('raffles.show', submission_id=sub_id))
+
+    user = _try_get_user_from_session()
+    raffle.queue(raffle_params=_raffle_params_from_form(form),
+                 user=user,
+                 job_id=sub_id)
+    return redirect(url_for('raffles.status', job_id=sub_id))
+
+
 def _filter_submissions(submissions_list):
     existing_raffle_ids = [tuple[0] for tuple in Raffle.query.
                            with_entities(Raffle.submission_id).all()]
     return [sub for sub in submissions_list if
             sub['id'] not in existing_raffle_ids]
+
+
+def _validate_raffle_form(form):
+    # Validate presence of required keys.
+    REQUIRED_KEYS = {'submissionUrl', 'winnerCount', 'minAge', 'minComment',
+                     'minLink'}
+    if not REQUIRED_KEYS.issubset(form.keys()):
+        return False
+
+    # Validate integer-value keys.
+    # All values must be non-negative. winnerCount must be at least 1.
+    INT_KEYS = {'minAge', 'winnerCount', 'minComment', 'minLink'}
+    for key in INT_KEYS:
+        val = form.get(key, type=int)
+        if (not isinstance(val, int)) or (val < 0) or \
+           (key == 'winnerCount' and (val < 1 or val > 25)):
+            return False
+
+    # Validate that the submission exists
+    url = _ensure_protocol(form.get('submissionUrl'))
+    if not reddit.get_submission(sub_url=url):
+        return False
+
+    return True
+
+
+def _ensure_protocol(url):
+    if url.startswith('http'):
+        return url
+    return 'https://' + url
+
+
+def _try_get_user_from_session():
+    if 'reddit_username' in session:
+        return User.query \
+                   .filter_by(username=session['reddit_username']) \
+                   .first()
+    else:
+        return None
+
+
+def _raffle_params_from_form(form):
+    return {
+        'submission_url': form.get('submissionUrl'),
+        'winner_count': form.get('winnerCount', type=int),
+        'min_account_age': form.get('minAge', type=int),
+        'min_comment_karma': form.get('minComment', type=int),
+        'min_link_karma': form.get('minLink', type=int)
+    }
+
+
+def _raffle_exists(sub_id):
+    return Raffle.query.filter_by(submission_id=sub_id).scalar()
