+++ tests/test_routes/test_api.py
@@ -1,5 +1,7 @@
 from flask import url_for, session
 from app.util import reddit
+from app.routes import api
+from app.jobs.raffle_job import raffle
 
 
 def test_submissions_no_token(client):
@@ -40,3 +42,52 @@ def test_submission_valid_submission(client, monkeypatch, db_session):
     monkeypatch.setattr(reddit, 'get_submission', lambda sub_url: 'something')
     res = client.get(url_for('api.submission', url='some_url'))
     assert res.status_code == 200
+
+
+def test_post_new_raffle_no_params(client):
+    assert client.post(url_for('api.new_raffle')).status_code == 422
+
+
+def test_post_new_raffle_invalid_params(client):
+    res = client.post(url_for('api.new_raffle'), data=_invalid_form_params())
+    assert res.status_code == 422
+
+
+def test_post_new_raffle_valid_params(client, monkeypatch):
+    monkeypatch.setattr(reddit, 'get_submission', lambda sub_url: True)
+    monkeypatch.setattr(api, '_raffle_exists', lambda sub_url: False)
+    monkeypatch.setattr(raffle, 'queue', _stub_raffle_job)
+    params = _valid_form_params()
+    res = client.post(url_for('api.new_raffle'), data=_valid_form_params())
+    assert res.status_code == 302
+    assert url_for('raffles.status', job_id='57xvjb') in \
+        res.headers['location']
+
+
+def test_post_new_raffle_valid_params_existing_raffle(client, monkeypatch):
+    monkeypatch.setattr(reddit, 'get_submission', lambda sub_url: True)
+    monkeypatch.setattr(api, '_raffle_exists', lambda sub_url: True)
+    monkeypatch.setattr(raffle, 'queue', _stub_raffle_job)
+    params = _valid_form_params()
+    res = client.post(url_for('api.new_raffle'), data=_valid_form_params())
+    assert res.status_code == 302
+    assert url_for('raffles.show', submission_id='57xvjb') in \
+        res.headers['location']
+
+
+def _valid_form_params():
+    return {
+        'submissionUrl': 'https://redd.it/57xvjb',
+        'winnerCount': 1,
+        'minAge': 0,
+        'minComment': 0,
+        'minLink': 0
+    }
+
+
+def _invalid_form_params():
+    return _valid_form_params().update({'winnerCount': 0})
+
+
+def _stub_raffle_job(raffle_params, user, job_id):
+    return [raffle_params, user, job_id]
